+++ fs/multifs.py
@@ -280,24 +280,24 @@ class MultiFS(FS):
         if not exists:
             raise errors.ResourceNotFound(path)
 
-    def getbytes(self, path):
+    def readbytes(self, path):
         # type: (Text) -> bytes
         self.check()
         fs = self._delegate(path)
         if fs is None:
             raise errors.ResourceNotFound(path)
-        return fs.getbytes(path)
+        return fs.readbytes(path)
 
-    def getfile(self, path, file, chunk_size=None, **options):
+    def download(self, path, file, chunk_size=None, **options):
         # type: (Text, BinaryIO, Optional[int], **Any) -> None
         fs = self._delegate_required(path)
-        return fs.getfile(path, file, chunk_size=chunk_size, **options)
+        return fs.download(path, file, chunk_size=chunk_size, **options)
 
-    def gettext(self, path, encoding=None, errors=None, newline=""):
+    def readtext(self, path, encoding=None, errors=None, newline=""):
         # type: (Text, Optional[Text], Optional[Text], Text) -> Text
         self.check()
         fs = self._delegate_required(path)
-        return fs.gettext(path, encoding=encoding, errors=errors, newline=newline)
+        return fs.readtext(path, encoding=encoding, errors=errors, newline=newline)
 
     def getsize(self, path):
         # type: (Text) -> int
@@ -400,15 +400,17 @@ class MultiFS(FS):
             **kwargs
         )
 
-    def setbinfile(self, path, file):
-        # type: (Text, BinaryIO) -> None
-        self._writable_required(path).setbinfile(path, file)
+    def upload(self, path, file, chunk_size=None, **options):
+        # type: (Text, BinaryIO, Optional[int], **Any) -> None
+        self._writable_required(path).upload(
+            path, file, chunk_size=chunk_size, **options
+        )
 
-    def setbytes(self, path, contents):
+    def writebytes(self, path, contents):
         # type: (Text, bytes) -> None
-        self._writable_required(path).setbytes(path, contents)
+        self._writable_required(path).writebytes(path, contents)
 
-    def settext(
+    def writetext(
         self,
         path,  # type: Text
         contents,  # type: Text
@@ -418,6 +420,6 @@ class MultiFS(FS):
     ):
         # type: (...) -> None
         write_fs = self._writable_required(path)
-        return write_fs.settext(
+        return write_fs.writetext(
             path, contents, encoding=encoding, errors=errors, newline=newline
         )